Rakudo Star 2012.01
Rakudo Star comes with its own version of Parrot that is designed to match the
target release of rakudo, so I am removing the Parrot dependency and having it
build against packages version of Parrot.

Also, having it install all of the tools, this includes parrot, nqp, and all
of the modules that come with Rakudo Star instead of just installing the perl6
executable.

Also added caveat about installing parrot and rakudo side by side.

Signed-off-by: Adam Vandenberg <flangy@gmail.com>