Don't override default -O flag.
Way back in the day, Homebrew defaulted to LLVM/-O3. A lot of stuff
failed to compile, and a lot of stuff just took forever with -O4.

We don't default this way anymore, and in fact use -Os per Apple
guidelines.

So remove these old "fixes" since they are no longer needed.