mapnik: fix compile error with system Python
Mapnik by default tries to install to the site-packages in /System
when brewed against system Python.  This causes a build error
because that directory is not writeable.  Mapnik correctly
determines the versioned site-packages name without modification,
but for Homebrew it needs the proper prefix.

* Add the standard `which_python` function to the formula.
* Add a scons argument for `PYTHON_PREFIX`
* Add the standard caveat for `PYTHONPATH`.

Discussed upstream: https://github.com/mapnik/mapnik/issues/1155
and tested on Lion against the system Python.

Fixes #11389

Closes #11407.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>