avidemux: adjust merged patch, use fails_with
Avidemux-2.5.6 had its one patch merged upstream the day this
formula was accepted.  That was hoped for but unexpected.  This
commit adjusts the conditional for def patches so that it only
gets applied to stable.  Fixes a patch failure for 32bit + HEAD.

Also in this commit, remove the hardcoded `ENV.llvm` and switch
to using `fails_with :clang`, surrounded by a conditional so it
can still be built with clang if `--with-debug`.

Closes #11959.

Signed-off-by: Jack Nagel <jacknagel@gmail.com>